+++ Engine/src/core/camera/Camera.h
@@ -1,38 +1,34 @@
 #pragma once
-
-#include <GLEW\glew.h>
-
-#include <glm\glm.hpp>
-#include <glm\gtc\matrix_transform.hpp>
+#include "En.h"
 
 enum Camera_Movement { FORWARD, BACKWARD, LEFT, RIGHT, STOP, ROLL_LEFT, ROLL_RIGHT};
 
-const GLfloat YAW = -90.0f;
-const GLfloat PITCH = 0.0f;
-const GLfloat SPEED = 0.0f;
-const GLfloat SENSITIVITY = 0.10f;
-const GLfloat FIELD_OF_VIEW = 45.0f;
+const float YAW = -90.0f;
+const float PITCH = 0.0f;
+const float SPEED = 0.0f;
+const float SENSITIVITY = 0.10f;
+const float FIELD_OF_VIEW = 45.0f;
 
 class Camera {
 public:
 	Camera(glm::vec3 position = glm::vec3(0.0f, 0.0f, -90.0f),
 		   glm::vec3 up = glm::vec3(0.0f, 1.0f, 0.0f),
-		   GLfloat yaw = YAW, GLfloat pitch = PITCH,
+		   float yaw = YAW, float pitch = PITCH,
 		   glm::vec3 front = glm::vec3(0.0f, 0.0f, -1.0f),
-		   GLfloat movement_speed = SPEED,
-		   GLfloat mouse_sensitiviy = SENSITIVITY,
-		   GLfloat field_of_view = FIELD_OF_VIEW);
+			float movement_speed = SPEED,
+			float mouse_sensitiviy = SENSITIVITY,
+			float field_of_view = FIELD_OF_VIEW);
 	~Camera();
 
-	void update(float delta_time);
+	void update(double delta_time);
 
 	void process_keyboard(Camera_Movement direction);
-	void process_mouse(GLfloat x_offset, GLfloat y_offset, bool constrain_pitch = true);
-	void process_mouse_scroll(GLfloat y_offset);
+	void process_mouse(double x_offset, double y_offset, bool constrain_pitch = true);
+	void process_mouse_scroll(double y_offset);
 
 	void update_camera_vectors();
 
-	GLfloat get_field_of_view();
+	float get_field_of_view();
 	glm::mat4 get_view_matrix();
 	glm::mat4 get_persp_proj_matrix();
 	glm::vec3 get_pos();
@@ -48,10 +44,10 @@ public:
 	void set_parent_model_mat(glm::mat4 parent_model_mat);
 	
 	/* Euler Angles */
-	GLfloat yaw = 0;
-	GLfloat pitch = 0;
-	GLfloat roll = 0;
-	GLfloat field_of_view;
+	float yaw = 0;
+	float pitch = 0;
+	float roll = 0;
+	float field_of_view;
 
 	bool first_person = false;
 
@@ -65,10 +61,10 @@ private:
 	glm::mat4 persp_proj_matrix;
 	glm::mat4 view_matrix;
 	/* Camera Options */
-	GLfloat movement_speed;
-	GLfloat acceleration = 0.4f;
-	GLfloat max_velocity = 0.6f;
-	GLfloat mouse_sensitivity;
+	float movement_speed;
+	float acceleration = 0.4f;
+	float max_velocity = 0.6f;
+	float mouse_sensitivity;
 	Camera_Movement direction;
 	
 	glm::mat4 parent_model_mat;
