+++ Engine/src/core/game_objects/lights/Light.cpp
@@ -1,7 +1,5 @@
 #include "Light.h"
 #include "..\..\Application.h"
-#include <glm\gtc\type_ptr.hpp>
-#include <glm\ext.hpp>
 
 Light::Light(Application* app, glm::vec3 light_position, glm::vec3 light_color) : GameObject("Light", app, new Model("models/sphere.obj")), light_position(light_position), light_color(light_color) {
 	this->set_model_mat(glm::translate(glm::mat4(1.0f), light_position));
@@ -37,7 +35,7 @@ void Light::render() {
 	shader_program->stop();
 }
 
-void Light::update(float delta_time) {
+void Light::update(double delta_time) {
 	GameObject::update(delta_time);
 	app->update_lights();
 }
