+++ Engine/src/core/game_objects/particle/ParticleMaster.h
@@ -7,7 +7,7 @@ public:
 	ParticleMaster(glm::mat4 proj_mat);
 	~ParticleMaster();
 
-	void update(float delta_time);
+	void update(double delta_time);
 	void render(Camera* camera);
 	void add_particle(Particle* p);
 	void apply_force(glm::vec3 force);
