+++ Real-Time-Animation_1/src/game_objects/plane/Plane.cpp
@@ -1,10 +1,4 @@
 #include "Plane.h"
-#include "../../Application.h"
-#include <glm\gtc\matrix_transform.hpp>
-#include <glm/ext.hpp>
-#include <glm/gtx/matrix_decompose.hpp>
-#include <tween\tween.h>
-
 
 Plane::Plane(std::string name, Application * app, Model* model, Camera* camera) : GameObject(name, app, model) {
 	camera = app->get_camera();
@@ -59,7 +53,7 @@ void Plane::attach_arrows(Application * app) {
 
 }
 
-void Plane::update(float delta_time) {
+void Plane::update(double delta_time) {
 	elapsed += delta_time;
 	tween::Tween::updateTweens(elapsed);
 
@@ -91,7 +85,7 @@ void Plane::update(float delta_time) {
 	if (app->get_camera()->first_person) {
 		app->get_camera()->set_parent_model_mat(get_model_mat());
 	}
-	
+
 	wheels->update(delta_time);
 	propeller->update(delta_time);
 }
@@ -153,15 +147,15 @@ void Plane::start_yaw() {
 	animation_start = true;
 }
 
-void Plane::update_quaternion_speed(float delta_time) {
+void Plane::update_quaternion_speed(double delta_time) {
 	/* Along Y Axis*/
-	glm::quat new_quat = glm::angleAxis(glm::radians(quaternion_speed.y * delta_time), axis_y);
+	glm::quat new_quat = glm::angleAxis(glm::radians(quaternion_speed.y * (float)delta_time), axis_y);
 	transform = new_quat * transform;
 
-	new_quat = glm::angleAxis(glm::radians(quaternion_speed.x *delta_time), axis_x);
+	new_quat = glm::angleAxis(glm::radians(quaternion_speed.x *(float)delta_time), axis_x);
 	transform = new_quat * transform;
 
-	new_quat = glm::angleAxis(glm::radians(quaternion_speed.z *delta_time), axis_z);
+	new_quat = glm::angleAxis(glm::radians(quaternion_speed.z *(float)delta_time), axis_z);
 	transform = new_quat * transform;
 
 	// Rotate on Y axis
@@ -186,7 +180,7 @@ void Plane::set_speed_z(float speed) {
 	this->quaternion_speed = glm::vec3(quaternion_speed.x, quaternion_speed.y, speed);
 }
 
-void Plane::set_toon_shading(bool toon_shading){
+void Plane::set_toon_shading(bool toon_shading) {
 	GameObject::set_toon_shading(toon_shading);
 	wheels->set_toon_shading(toon_shading);
 	propeller->set_toon_shading(toon_shading);
